+++ game-net/src/main/java/com/wjybxx/fastjgame/net/binary/ProtoEnumCodec.java
@@ -33,7 +33,7 @@ public class ProtoEnumCodec<T extends ProtocolMessageEnum> extends PojoCodec<T> 
     private final Class<T> enumClass;
     private final Internal.EnumLiteMap<T> mapper;
 
-    public ProtoEnumCodec(int providerId, int classId, Class<T> enumClass, Internal.EnumLiteMap<T> mapper) {
+    public ProtoEnumCodec(byte providerId, int classId, Class<T> enumClass, Internal.EnumLiteMap<T> mapper) {
         super(providerId, classId);
         this.enumClass = enumClass;
         this.mapper = mapper;
