+++ game-net/src/main/java/com/wjybxx/fastjgame/net/binary/ProtoMessageCodec.java
@@ -34,7 +34,7 @@ public class ProtoMessageCodec<T extends AbstractMessage> extends PojoCodec<T> {
     private final Class<T> messageClass;
     private final Parser<T> parser;
 
-    public ProtoMessageCodec(int providerId, int classId, Class<T> messageClass, Parser<T> parser) {
+    public ProtoMessageCodec(byte providerId, int classId, Class<T> messageClass, Parser<T> parser) {
         super(providerId, classId);
         this.messageClass = messageClass;
         this.parser = parser;
